Move base event classes into private module.
